move bulk of rag into its util.py
