+++ example/rag/config.py
@@ -7,9 +7,9 @@ milvus_uri = "file:/milvus.db"
 
 embedding_model_id = "all-MiniLM-L6-v2"
 
-model_id = "ollama:llama3.2"
-#model_id = "ollama:granite3.2"
 #model_id = "openai:gpt-4o"
+#model_id = "ollama:granite3.2"
+model_id = "ollama:llama3.2"
 
 url = 'https://www.gutenberg.org/cache/epub/8714/pg8714.txt'
 data_file = Path("four_plays_of_aeschylus.txt")
